fixup cyclic dependency by timestamp and object module
