+++ crates/rooch-types/src/framework/mod.rs
@@ -23,6 +23,7 @@ pub mod genesis;
 pub mod native_validator;
 pub mod onchain_config;
 pub mod session_key;
+pub mod timestamp;
 pub mod transaction_validator;
 pub mod transfer;
