+++ frameworks/moveos-stdlib/doc/README.md
@@ -43,6 +43,7 @@ This is the reference documentation of the MoveOS standard library.
 -  [`0x2::string_utils`](string_utils.md#0x2_string_utils)
 -  [`0x2::table`](table.md#0x2_table)
 -  [`0x2::table_vec`](table_vec.md#0x2_table_vec)
+-  [`0x2::timestamp`](timestamp.md#0x2_timestamp)
 -  [`0x2::tx_context`](tx_context.md#0x2_tx_context)
 -  [`0x2::tx_meta`](tx_meta.md#0x2_tx_meta)
 -  [`0x2::tx_result`](tx_result.md#0x2_tx_result)
