+++ frameworks/moveos-stdlib/sources/timestamp.move
@@ -7,17 +7,10 @@
 /// * L1 block: update the timestamp via L1s block header timestamp
 /// * L2 transactions: update the timestamp via L2 transaction's timestamp 
 module moveos_std::timestamp {
-
+    use moveos_std::object::{update_global_time, try_update_global_time_internal, Timestamp};
     use moveos_std::object;
-    use moveos_std::signer;
     use moveos_std::core_addresses;
-
-    friend moveos_std::genesis;
-
-    /// A object holding the current Unix time in milliseconds
-    struct Timestamp has key {
-        milliseconds: u64,
-    }
+    use moveos_std::signer;
 
     /// Conversion factor between seconds and milliseconds
     const MILLI_CONVERSION_FACTOR: u64 = 1000;
@@ -25,20 +18,6 @@ module moveos_std::timestamp {
     /// An invalid timestamp was provided
     const ErrorInvalidTimestamp: u64 = 1;
     const ErrorNotGenesisAddress: u64 = 2;
-
-    public(friend) fun genesis_init(_genesis_account: &signer, initial_time_milliseconds: u64) {
-        let timestamp = Timestamp { milliseconds: initial_time_milliseconds };
-        let obj = object::new_named_object(timestamp);
-        object::transfer_extend(obj, @moveos_std);
-    }
-
-    /// Updates the global clock time, if the new time is smaller than the current time, aborts.
-    public(friend) fun update_global_time(timestamp_milliseconds: u64) {
-        let current_timestamp = timestamp_mut(); 
-        let now = current_timestamp.milliseconds;
-        assert!(now < timestamp_milliseconds, ErrorInvalidTimestamp);
-        current_timestamp.milliseconds = timestamp_milliseconds;
-    }
 
     /// Tries to update the global clock time, if the new time is smaller than the current time, ignores the update, and returns false.
     /// Only the framework genesis account can update the global clock time.
@@ -48,46 +27,26 @@ module moveos_std::timestamp {
         try_update_global_time_internal(timestamp_milliseconds)
     }
 
-    public(friend) fun try_update_global_time_internal(timestamp_milliseconds: u64) : bool {
-        let current_timestamp = timestamp_mut(); 
-        let now = current_timestamp.milliseconds;
-        if(now < timestamp_milliseconds) {
-            current_timestamp.milliseconds = timestamp_milliseconds;
-            true
-        }else{
-            false
-        }
-    }
-
-    fun timestamp_mut(): &mut Timestamp {
-        let object_id = object::named_object_id<Timestamp>();
-        let obj = object::borrow_mut_object_extend<Timestamp>(object_id);
-        object::borrow_mut(obj)
-    }
-
     public fun timestamp(): &Timestamp {
-        let object_id = object::named_object_id<Timestamp>();
-        let obj = object::borrow_object<Timestamp>(object_id);
-        object::borrow(obj)
+        object::timestamp()
     }
 
     public fun milliseconds(self: &Timestamp): u64 {
-        self.milliseconds
+        object::milliseconds(self)
     }
 
     public fun seconds(self: &Timestamp): u64 {
-        self.milliseconds / MILLI_CONVERSION_FACTOR
+        object::seconds(self)
     }
 
     /// Gets the current time in milliseconds.
     public fun now_milliseconds(): u64 {
-        let timestamp = timestamp();
-        timestamp.milliseconds
+        object::now_milliseconds()
     }
 
     /// Gets the current time in seconds.
     public fun now_seconds(): u64 {
-        now_milliseconds() / MILLI_CONVERSION_FACTOR
+        object::now_seconds()
     }
 
     public fun seconds_to_milliseconds(seconds: u64): u64 {
@@ -95,8 +54,15 @@ module moveos_std::timestamp {
     }
 
     fun fast_forward_seconds(timestamp_seconds: u64) {
-        let now_milliseconds = now_milliseconds();
+        let timestamp = timestamp();
+        let now_milliseconds = milliseconds(timestamp);
         update_global_time(now_milliseconds + (timestamp_seconds * MILLI_CONVERSION_FACTOR));
+    }
+
+    public fun fast_forward_seconds_by_system(genesis_account: &signer, timestamp_seconds: u64) {
+        let genesis_address = signer::address_of(genesis_account);
+        assert!(core_addresses::is_system_reserved_address(genesis_address), ErrorNotGenesisAddress);
+        fast_forward_seconds(timestamp_seconds)
     }
 
     #[test_only]
@@ -112,11 +78,5 @@ module moveos_std::timestamp {
     #[test_only]
     public fun fast_forward_seconds_for_test(timestamp_seconds: u64) {
         fast_forward_seconds(timestamp_seconds)
-    }
-
-    /// Fast forwards the clock by the given number of seconds, but only if the chain is in local mode.
-    public entry fun fast_forward_seconds_for_local(timestamp_seconds: u64) {
-        assert!(rooch_framework::chain_id::is_local(), ErrorInvalidTimestamp);
-        fast_forward_seconds(timestamp_seconds);
     }
 }
