+++ frameworks/rooch-framework/doc/transaction_validator.md
@@ -14,6 +14,7 @@
 <b>use</b> <a href="">0x2::account</a>;
 <b>use</b> <a href="">0x2::gas_schedule</a>;
 <b>use</b> <a href="">0x2::signer</a>;
+<b>use</b> <a href="">0x2::timestamp</a>;
 <b>use</b> <a href="">0x2::tx_context</a>;
 <b>use</b> <a href="">0x2::tx_result</a>;
 <b>use</b> <a href="account.md#0x3_account">0x3::account</a>;
@@ -27,7 +28,6 @@
 <b>use</b> <a href="gas_coin.md#0x3_gas_coin">0x3::gas_coin</a>;
 <b>use</b> <a href="multichain_address.md#0x3_multichain_address">0x3::multichain_address</a>;
 <b>use</b> <a href="session_key.md#0x3_session_key">0x3::session_key</a>;
-<b>use</b> <a href="timestamp.md#0x3_timestamp">0x3::timestamp</a>;
 <b>use</b> <a href="transaction.md#0x3_transaction">0x3::transaction</a>;
 <b>use</b> <a href="transaction_fee.md#0x3_transaction_fee">0x3::transaction_fee</a>;
 </code></pre>
